model/client/requirements.txt
gradio>=4.26.0
torch
torchvision
timm
pandas
pillow
numpy
python-dotenv
matplotlib
